+++ src/main/java/com/xxxmkxxx/customgui/client/hierarchy/window/frame/Frame.java
@@ -3,6 +3,6 @@ package com.xxxmkxxx.customgui.client.hierarchy.window.frame;
 import com.xxxmkxxx.customgui.client.hierarchy.window.position.Pos;
 
 public interface Frame {
-    boolean checkPosBelongs(int xPos, int yPos);
+    boolean checkPosBelongs(float xPos, float yPos);
     boolean checkPosBelongs(Pos pos);
 }
